Add a script to update all the plugins
Plugins are listed in bundle/plugins.
The script runs through each plugin and clones the repo if doesn't exist
or pulls if it does.
The script assumes all plugins are git repos,
and that they do not require authentication to clone/pull.
Many more assumptions are made - the script is not intended to be foolproof;
it is only intended to run properly if the plugin directories and list is kept
in the form it expects.